Add version switch logic to OpenAPI importer
